+++ matlab/calc_sum.h
@@ -1,4 +1,7 @@
-
 #define DLL_EXPORT __declspec(dllexport)
 
-DLL_EXPORT float my_func(float num1, float num2);
+DLL_EXPORT int my_func(int num1, int num2);
+
+
+
+
